+++ src/ext/BalExtension/wixstdba/WixStandardBootstrapperApplication.cpp
@@ -24,10 +24,12 @@ static const DWORD WIXSTDBA_ACQUIRE_PERCENTAGE = 30;
 
 static const LPCWSTR WIXSTDBA_VARIABLE_BUNDLE_FILE_VERSION = L"WixBundleFileVersion";
 static const LPCWSTR WIXSTDBA_VARIABLE_LANGUAGE_ID = L"WixStdBALanguageId";
+static const LPCWSTR WIXSTDBA_VARIABLE_RESTART_REQUIRED = L"WixStdBARestartRequired";
+static const LPCWSTR WIXSTDBA_VARIABLE_SHOW_VERSION = L"WixStdBAShowVersion";
+static const LPCWSTR WIXSTDBA_VARIABLE_SUPPRESS_OPTIONS_UI = L"WixStdBASuppressOptionsUI";
 
 enum WIXSTDBA_STATE
 {
-    WIXSTDBA_STATE_OPTIONS,
     WIXSTDBA_STATE_INITIALIZING,
     WIXSTDBA_STATE_INITIALIZED,
     WIXSTDBA_STATE_HELP,
@@ -60,7 +62,6 @@ enum WIXSTDBA_PAGE
     WIXSTDBA_PAGE_LOADING,
     WIXSTDBA_PAGE_HELP,
     WIXSTDBA_PAGE_INSTALL,
-    WIXSTDBA_PAGE_OPTIONS,
     WIXSTDBA_PAGE_MODIFY,
     WIXSTDBA_PAGE_PROGRESS,
     WIXSTDBA_PAGE_PROGRESS_PASSIVE,
@@ -74,7 +75,6 @@ static LPCWSTR vrgwzPageNames[] = {
     L"Loading",
     L"Help",
     L"Install",
-    L"Options",
     L"Modify",
     L"Progress",
     L"ProgressPassive",
@@ -84,32 +84,15 @@ static LPCWSTR vrgwzPageNames[] = {
 
 enum WIXSTDBA_CONTROL
 {
-    // Non-paged controls
-    WIXSTDBA_CONTROL_CLOSE_BUTTON = THEME_FIRST_ASSIGN_CONTROL_ID,
-    WIXSTDBA_CONTROL_MINIMIZE_BUTTON,
-
-    // Help page
-    WIXSTDBA_CONTROL_HELP_CANCEL_BUTTON,
-
     // Welcome page
-    WIXSTDBA_CONTROL_INSTALL_BUTTON,
-    WIXSTDBA_CONTROL_OPTIONS_BUTTON,
+    WIXSTDBA_CONTROL_INSTALL_BUTTON = THEME_FIRST_ASSIGN_CONTROL_ID,
     WIXSTDBA_CONTROL_EULA_RICHEDIT,
     WIXSTDBA_CONTROL_EULA_LINK,
     WIXSTDBA_CONTROL_EULA_ACCEPT_CHECKBOX,
-    WIXSTDBA_CONTROL_WELCOME_CANCEL_BUTTON,
-    WIXSTDBA_CONTROL_VERSION_LABEL,
-
-    // Options page
-    WIXSTDBA_CONTROL_FOLDER_EDITBOX,
-    WIXSTDBA_CONTROL_BROWSE_BUTTON,
-    WIXSTDBA_CONTROL_OK_BUTTON,
-    WIXSTDBA_CONTROL_CANCEL_BUTTON,
 
     // Modify page
     WIXSTDBA_CONTROL_REPAIR_BUTTON,
     WIXSTDBA_CONTROL_UNINSTALL_BUTTON,
-    WIXSTDBA_CONTROL_MODIFY_CANCEL_BUTTON,
 
     // Progress page
     WIXSTDBA_CONTROL_CACHE_PROGRESS_PACKAGE_TEXT,
@@ -130,40 +113,22 @@ enum WIXSTDBA_CONTROL
 
     // Success page
     WIXSTDBA_CONTROL_LAUNCH_BUTTON,
-    WIXSTDBA_CONTROL_SUCCESS_RESTART_TEXT,
     WIXSTDBA_CONTROL_SUCCESS_RESTART_BUTTON,
-    WIXSTDBA_CONTROL_SUCCESS_CANCEL_BUTTON,
 
     // Failure page
     WIXSTDBA_CONTROL_FAILURE_LOGFILE_LINK,
     WIXSTDBA_CONTROL_FAILURE_MESSAGE_TEXT,
-    WIXSTDBA_CONTROL_FAILURE_RESTART_TEXT,
     WIXSTDBA_CONTROL_FAILURE_RESTART_BUTTON,
-    WIXSTDBA_CONTROL_FAILURE_CANCEL_BUTTON,
 };
 
 static THEME_ASSIGN_CONTROL_ID vrgInitControls[] = {
-    { WIXSTDBA_CONTROL_CLOSE_BUTTON, L"CloseButton" },
-    { WIXSTDBA_CONTROL_MINIMIZE_BUTTON, L"MinimizeButton" },
-
-    { WIXSTDBA_CONTROL_HELP_CANCEL_BUTTON, L"HelpCancelButton" },
-
     { WIXSTDBA_CONTROL_INSTALL_BUTTON, L"InstallButton" },
-    { WIXSTDBA_CONTROL_OPTIONS_BUTTON, L"OptionsButton" },
     { WIXSTDBA_CONTROL_EULA_RICHEDIT, L"EulaRichedit" },
     { WIXSTDBA_CONTROL_EULA_LINK, L"EulaHyperlink" },
     { WIXSTDBA_CONTROL_EULA_ACCEPT_CHECKBOX, L"EulaAcceptCheckbox" },
-    { WIXSTDBA_CONTROL_WELCOME_CANCEL_BUTTON, L"WelcomeCancelButton" },
-    { WIXSTDBA_CONTROL_VERSION_LABEL, L"InstallVersion" },
-
-    { WIXSTDBA_CONTROL_FOLDER_EDITBOX, L"FolderEditbox" },
-    { WIXSTDBA_CONTROL_BROWSE_BUTTON, L"BrowseButton" },
-    { WIXSTDBA_CONTROL_OK_BUTTON, L"OptionsOkButton" },
-    { WIXSTDBA_CONTROL_CANCEL_BUTTON, L"OptionsCancelButton" },
 
     { WIXSTDBA_CONTROL_REPAIR_BUTTON, L"RepairButton" },
     { WIXSTDBA_CONTROL_UNINSTALL_BUTTON, L"UninstallButton" },
-    { WIXSTDBA_CONTROL_MODIFY_CANCEL_BUTTON, L"ModifyCancelButton" },
 
     { WIXSTDBA_CONTROL_CACHE_PROGRESS_PACKAGE_TEXT, L"CacheProgressPackageText" },
     { WIXSTDBA_CONTROL_CACHE_PROGRESS_BAR, L"CacheProgressbar" },
@@ -179,15 +144,11 @@ static THEME_ASSIGN_CONTROL_ID vrgInitControls[] = {
     { WIXSTDBA_CONTROL_PROGRESS_CANCEL_BUTTON, L"ProgressCancelButton" },
 
     { WIXSTDBA_CONTROL_LAUNCH_BUTTON, L"LaunchButton" },
-    { WIXSTDBA_CONTROL_SUCCESS_RESTART_TEXT, L"SuccessRestartText" },
     { WIXSTDBA_CONTROL_SUCCESS_RESTART_BUTTON, L"SuccessRestartButton" },
-    { WIXSTDBA_CONTROL_SUCCESS_CANCEL_BUTTON, L"SuccessCancelButton" },
 
     { WIXSTDBA_CONTROL_FAILURE_LOGFILE_LINK, L"FailureLogFileLink" },
     { WIXSTDBA_CONTROL_FAILURE_MESSAGE_TEXT, L"FailureMessageText" },
-    { WIXSTDBA_CONTROL_FAILURE_RESTART_TEXT, L"FailureRestartText" },
     { WIXSTDBA_CONTROL_FAILURE_RESTART_BUTTON, L"FailureRestartButton" },
-    { WIXSTDBA_CONTROL_FAILURE_CANCEL_BUTTON, L"FailureCloseButton" },
 };
 
 typedef struct _WIXSTDBA_PREREQ_PACKAGE
@@ -886,6 +847,8 @@ public: // IBootstrapperApplication
 
         // If a restart was encountered and we are not suppressing restarts, then restart is required.
         m_fRestartRequired = (BOOTSTRAPPER_APPLY_RESTART_NONE != restart && BOOTSTRAPPER_RESTART_NEVER < m_command.restart);
+        BalSetStringVariable(WIXSTDBA_VARIABLE_RESTART_REQUIRED, m_fRestartRequired ? L"1" : NULL);
+
         // If a restart is required and we're not displaying a UI or we are not supposed to prompt for restart then allow the restart.
         m_fAllowRestart = m_fRestartRequired && (BOOTSTRAPPER_DISPLAY_FULL > m_command.display || BOOTSTRAPPER_RESTART_PROMPT < m_command.restart);
 
@@ -1085,7 +1048,7 @@ private: // privates
         }
         else
         {
-            hr = ParseBootrapperApplicationDataFromXml(pixdManifest);
+            hr = ParseBootstrapperApplicationDataFromXml(pixdManifest);
             BalExitOnFailure(hr, "Failed to read bootstrapper application data.");
         }
 
@@ -1399,7 +1362,7 @@ private: // privates
     }
 
 
-    HRESULT ParseBootrapperApplicationDataFromXml(
+    HRESULT ParseBootstrapperApplicationDataFromXml(
         __in IXMLDOMDocument* pixdManifest
         )
     {
@@ -1442,9 +1405,10 @@ private: // privates
         {
             hr = S_OK;
         }
-        else if (SUCCEEDED(hr))
+        else if (SUCCEEDED(hr) && dwBool)
         {
-            m_fSuppressOptionsUI = 0 < dwBool;
+            hr = BalSetNumericVariable(WIXSTDBA_VARIABLE_SUPPRESS_OPTIONS_UI, 1);
+            BalExitOnFailure(hr, "Failed to set '%ls' variable.", WIXSTDBA_VARIABLE_SUPPRESS_OPTIONS_UI);
         }
         BalExitOnFailure(hr, "Failed to get SuppressOptionsUI value.");
 
@@ -1477,9 +1441,10 @@ private: // privates
         {
             hr = S_OK;
         }
-        else if (SUCCEEDED(hr))
+        else if (SUCCEEDED(hr) && dwBool)
         {
-            m_fShowVersion = 0 < dwBool;
+            hr = BalSetNumericVariable(WIXSTDBA_VARIABLE_SHOW_VERSION, 1);
+            BalExitOnFailure(hr, "Failed to set '%ls' variable.", WIXSTDBA_VARIABLE_SHOW_VERSION);
         }
         BalExitOnFailure(hr, "Failed to get ShowVersion value.");
 
@@ -1752,22 +1717,6 @@ private: // privates
                     pBA->OnClickAcceptCheckbox();
                     return 0;
 
-                case WIXSTDBA_CONTROL_OPTIONS_BUTTON:
-                    pBA->OnClickOptionsButton();
-                    return 0;
-
-                case WIXSTDBA_CONTROL_BROWSE_BUTTON:
-                    pBA->OnClickOptionsBrowseButton();
-                    return 0;
-
-                case WIXSTDBA_CONTROL_OK_BUTTON:
-                    pBA->OnClickOptionsOkButton();
-                    return 0;
-
-                case WIXSTDBA_CONTROL_CANCEL_BUTTON:
-                    pBA->OnClickOptionsCancelButton();
-                    return 0;
-
                 case WIXSTDBA_CONTROL_INSTALL_BUTTON:
                     pBA->OnClickInstallButton();
                     return 0;
@@ -1789,13 +1738,7 @@ private: // privates
                     pBA->OnClickRestartButton();
                     return 0;
 
-                case WIXSTDBA_CONTROL_HELP_CANCEL_BUTTON: __fallthrough;
-                case WIXSTDBA_CONTROL_WELCOME_CANCEL_BUTTON: __fallthrough;
-                case WIXSTDBA_CONTROL_MODIFY_CANCEL_BUTTON: __fallthrough;
-                case WIXSTDBA_CONTROL_PROGRESS_CANCEL_BUTTON: __fallthrough;
-                case WIXSTDBA_CONTROL_SUCCESS_CANCEL_BUTTON: __fallthrough;
-                case WIXSTDBA_CONTROL_FAILURE_CANCEL_BUTTON: __fallthrough;
-                case WIXSTDBA_CONTROL_CLOSE_BUTTON:
+                case WIXSTDBA_CONTROL_PROGRESS_CANCEL_BUTTON:
                     pBA->OnClickCloseButton();
                     return 0;
                 }
@@ -2122,31 +2065,10 @@ private: // privates
 
                     BOOL fAcceptedLicense = !ThemeControlExists(m_pTheme, WIXSTDBA_CONTROL_EULA_ACCEPT_CHECKBOX) || !ThemeControlEnabled(m_pTheme, WIXSTDBA_CONTROL_EULA_ACCEPT_CHECKBOX) || ThemeIsControlChecked(m_pTheme, WIXSTDBA_CONTROL_EULA_ACCEPT_CHECKBOX);
                     ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_INSTALL_BUTTON, fAcceptedLicense);
-
-                    // If there is an "Options" page, the "Options" button exists, and it hasn't been suppressed, then enable the button.
-                    BOOL fOptionsEnabled = m_rgdwPageIds[WIXSTDBA_PAGE_OPTIONS] && ThemeControlExists(m_pTheme, WIXSTDBA_CONTROL_OPTIONS_BUTTON) && !m_fSuppressOptionsUI;
-                    ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_OPTIONS_BUTTON, fOptionsEnabled);
-
-                    // Show/Hide the version label if it exists.
-                    if (m_rgdwPageIds[WIXSTDBA_PAGE_OPTIONS] && ThemeControlExists(m_pTheme, WIXSTDBA_CONTROL_VERSION_LABEL) && !m_fShowVersion)
-                    {
-                        ThemeShowControl(m_pTheme, WIXSTDBA_CONTROL_VERSION_LABEL, SW_HIDE);
-                    }
                 }
                 else if (m_rgdwPageIds[WIXSTDBA_PAGE_MODIFY] == dwNewPageId)
                 {
                     ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_REPAIR_BUTTON, !m_fSuppressRepair);
-                }
-                else if (m_rgdwPageIds[WIXSTDBA_PAGE_OPTIONS] == dwNewPageId)
-                {
-                    HRESULT hr = BalGetStringVariable(WIXSTDBA_VARIABLE_INSTALL_FOLDER, &sczUnformattedText);
-                    if (SUCCEEDED(hr))
-                    {
-                        // If the wix developer is showing a hidden variable in the UI, then obviously they don't care about keeping it safe
-                        // so don't go down the rabbit hole of making sure that this is securely freed.
-                        BalFormatString(sczUnformattedText, &sczText);
-                        ThemeSetTextControl(m_pTheme, WIXSTDBA_CONTROL_FOLDER_EDITBOX, sczText);
-                    }
                 }
                 else if (m_rgdwPageIds[WIXSTDBA_PAGE_SUCCESS] == dwNewPageId) // on the "Success" page, check if the restart or launch button should be enabled.
                 {
@@ -2165,7 +2087,6 @@ private: // privates
                     }
 
                     ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_LAUNCH_BUTTON, fLaunchTargetExists && BOOTSTRAPPER_ACTION_UNINSTALL < m_plannedAction);
-                    ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_SUCCESS_RESTART_TEXT, fShowRestartButton);
                     ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_SUCCESS_RESTART_BUTTON, fShowRestartButton);
                 }
                 else if (m_rgdwPageIds[WIXSTDBA_PAGE_FAILURE] == dwNewPageId) // on the "Failure" page, show error message and check if the restart button should be enabled.
@@ -2240,11 +2161,10 @@ private: // privates
 
                     ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_FAILURE_LOGFILE_LINK, fShowLogLink);
                     ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_FAILURE_MESSAGE_TEXT, fShowErrorMessage);
-                    ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_FAILURE_RESTART_TEXT, fShowRestartButton);
                     ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_FAILURE_RESTART_BUTTON, fShowRestartButton);
                 }
 
-                HRESULT hr = ThemeShowPageEx(m_pTheme, dwOldPageId, SW_HIDE, m_fStateChangeIsCancel ? THEME_SHOW_PAGE_REASON_CANCEL : THEME_SHOW_PAGE_REASON_DEFAULT);
+                HRESULT hr = ThemeShowPage(m_pTheme, dwOldPageId, SW_HIDE);
                 if (FAILED(hr))
                 {
                     BalLogError(hr, "Failed to hide page: %u", dwOldPageId);
@@ -2322,77 +2242,6 @@ private: // privates
     {
         BOOL fAcceptedLicense = ThemeIsControlChecked(m_pTheme, WIXSTDBA_CONTROL_EULA_ACCEPT_CHECKBOX);
         ThemeControlEnable(m_pTheme, WIXSTDBA_CONTROL_INSTALL_BUTTON, fAcceptedLicense);
-    }
-
-
-    //
-    // OnClickOptionsButton - show the options page.
-    //
-    void OnClickOptionsButton()
-    {
-        m_stateBeforeOptions = m_state;
-        SetState(WIXSTDBA_STATE_OPTIONS, S_OK);
-    }
-
-
-    //
-    // OnClickOptionsBrowseButton - browse for install folder on the options page.
-    //
-    void OnClickOptionsBrowseButton()
-    {
-        WCHAR wzPath[MAX_PATH] = { };
-        BROWSEINFOW browseInfo = { };
-        PIDLIST_ABSOLUTE pidl = NULL;
-
-        browseInfo.hwndOwner = m_hWnd;
-        browseInfo.pszDisplayName = wzPath;
-        browseInfo.lpszTitle = m_pTheme->sczCaption;
-        browseInfo.ulFlags = BIF_RETURNONLYFSDIRS | BIF_USENEWUI;
-        pidl = ::SHBrowseForFolderW(&browseInfo);
-        if (pidl && ::SHGetPathFromIDListW(pidl, wzPath))
-        {
-            ThemeSetTextControl(m_pTheme, WIXSTDBA_CONTROL_FOLDER_EDITBOX, wzPath);
-        }
-
-        if (pidl)
-        {
-            ::CoTaskMemFree(pidl);
-        }
-
-        return;
-    }
-
-    //
-    // OnClickOptionsOkButton - accept the changes made by the options page.
-    //
-    void OnClickOptionsOkButton()
-    {
-        HRESULT hr = S_OK;
-        LPWSTR sczPath = NULL;
-
-        if (ThemeControlExists(m_pTheme, WIXSTDBA_CONTROL_FOLDER_EDITBOX))
-        {
-            hr = ThemeGetTextControl(m_pTheme, WIXSTDBA_CONTROL_FOLDER_EDITBOX, &sczPath);
-            ExitOnFailure(hr, "Failed to get text from folder edit box.");
-
-            // TODO: verify the path is valid.
-
-            hr = m_pEngine->SetVariableString(WIXSTDBA_VARIABLE_INSTALL_FOLDER, sczPath);
-            ExitOnFailure(hr, "Failed to set the install folder.");
-        }
-
-    LExit:
-        SetState(m_stateBeforeOptions, S_OK);
-        return;
-    }
-
-
-    //
-    // OnClickOptionsCancelButton - discard the changes made by the options page.
-    //
-    void OnClickOptionsCancelButton()
-    {
-        SetStateWithCancel(m_stateBeforeOptions, S_OK, TRUE);
     }
 
 
@@ -2599,19 +2448,6 @@ private: // privates
         __in HRESULT hrStatus
         )
     {
-        SetStateWithCancel(state, hrStatus, FALSE);
-    }
-
-
-    //
-    // SetStateWithCancel
-    //
-    void SetStateWithCancel(
-        __in WIXSTDBA_STATE state,
-        __in HRESULT hrStatus,
-        __in BOOL fIsCancel
-        )
-    {
         if (FAILED(hrStatus))
         {
             m_hrFinal = hrStatus;
@@ -2622,9 +2458,8 @@ private: // privates
             state = WIXSTDBA_STATE_FAILED;
         }
 
-        if (WIXSTDBA_STATE_OPTIONS == state || m_state < state)
+        if (m_state < state)
         {
-            m_fStateChangeIsCancel = fIsCancel;
             ::PostMessageW(m_hWnd, WM_WIXSTDBA_CHANGE_STATE, 0, state);
         }
     }
@@ -2700,10 +2535,6 @@ private: // privates
                     *pdwPageId = m_rgdwPageIds[WIXSTDBA_PAGE_MODIFY];
                     break;
                 }
-                break;
-
-            case WIXSTDBA_STATE_OPTIONS:
-                *pdwPageId = m_rgdwPageIds[WIXSTDBA_PAGE_OPTIONS];
                 break;
 
             case WIXSTDBA_STATE_PLANNING: __fallthrough;
@@ -2920,10 +2751,8 @@ public:
 
         m_sczLicenseFile = NULL;
         m_sczLicenseUrl = NULL;
-        m_fSuppressOptionsUI = FALSE;
         m_fSuppressDowngradeFailure = FALSE;
         m_fSuppressRepair = FALSE;
-        m_fShowVersion = FALSE;
 
         m_sdOverridableVariables = NULL;
         m_shPrereqSupportPackages = NULL;
@@ -3000,8 +2829,6 @@ private:
     HWND m_hWnd;
 
     WIXSTDBA_STATE m_state;
-    WIXSTDBA_STATE m_stateBeforeOptions;
-    BOOL m_fStateChangeIsCancel;
     HRESULT m_hrFinal;
 
     BOOL m_fStartedExecution;
@@ -3015,10 +2842,8 @@ private:
 
     LPWSTR m_sczLicenseFile;
     LPWSTR m_sczLicenseUrl;
-    BOOL m_fSuppressOptionsUI;
     BOOL m_fSuppressDowngradeFailure;
     BOOL m_fSuppressRepair;
-    BOOL m_fShowVersion;
 
     STRINGDICT_HANDLE m_sdOverridableVariables;